fixing missing endpoints
