+++ plugins/traction_innkeeper/traction_innkeeper/v1_0/innkeeper/models.py
@@ -320,23 +320,6 @@ class TenantRecord(BaseRecord):
             raise StorageNotFoundError("No TenantRecord found for the given wallet_id")
         return result[0]
 
-    @classmethod
-    async def retrieve_by_tenant_id(
-        cls,
-        session: ProfileSession,
-        record_id: str,
-        *,
-        for_update=False,
-    ) -> "TenantRecord":
-        """Retrieve TenantRecord by tenant_id.
-        Args:
-            session: the profile session to use
-            record_id: the tenant_id by which to filter
-        """
-        tenant_id = cls.transform_tenant_id(record_id)
-        record = await cls.retrieve_by_id(session, tenant_id, for_update=for_update)
-        return record
-
 
 class TenantRecordSchema(BaseRecordSchema):
     """Innkeeper Tenant Record Schema."""
