+++ plugins/traction_innkeeper/traction_innkeeper/v1_0/innkeeper/tenant_manager.py
@@ -7,9 +7,9 @@ from aries_cloudagent.core.error import BaseError
 from aries_cloudagent.core.profile import Profile
 from aries_cloudagent.messaging.models.base import BaseModelError
 from aries_cloudagent.multitenant.base import BaseMultitenantManager
-from aries_cloudagent.ledger.multiple_ledger.base_manager import (
-    BaseMultipleLedgerManager,
-)
+# from aries_cloudagent.ledger.multiple_ledger.base_manager import (
+#     BaseMultipleLedgerManager,
+# )
 from aries_cloudagent.storage.error import StorageError, StorageNotFoundError
 from aries_cloudagent.wallet.models.wallet_record import WalletRecord
